+++ apps/studio/data/database-cron-jobs/database-cron-jobs-infinite-query.ts
@@ -1,8 +1,8 @@
+import { getCronJobsSql } from '@supabase/pg-meta'
 import { InfiniteData, useInfiniteQuery } from '@tanstack/react-query'
 import { COST_THRESHOLD_ERROR, executeSql } from 'data/sql/execute-sql-query'
 import type { ResponseError, UseCustomInfiniteQueryOptions } from 'types'
 
-import { getCronJobsSql } from './database-cron-jobs.sql'
 import { databaseCronJobsKeys } from './keys'
 
 export const CRON_JOBS_PAGE_LIMIT = 20
