+++ apps/studio/data/database-cron-jobs/database-cron-jobs-minimal-infinite-query.ts
@@ -1,3 +1,4 @@
+import { getCronJobsMinimalSql } from '@supabase/pg-meta'
 import { InfiniteData, useInfiniteQuery } from '@tanstack/react-query'
 import { executeSql } from 'data/sql/execute-sql-query'
 import type { ResponseError, UseCustomInfiniteQueryOptions } from 'types'
@@ -7,7 +8,6 @@ import {
   CronJob,
   DatabaseCronJobRunsVariables,
 } from './database-cron-jobs-infinite-query'
-import { getCronJobsMinimalSql } from './database-cron-jobs.sql'
 import { databaseCronJobsKeys } from './keys'
 
 export async function getDatabaseCronJobsMinimal({
