+++ apps/studio/data/database-cron-jobs/schedule-clean-up-mutation.ts
@@ -1,12 +1,10 @@
+import { getScheduleDeleteCronJobRunDetailsSql } from '@supabase/pg-meta'
 import { useMutation } from '@tanstack/react-query'
 import { executeSql } from 'data/sql/execute-sql-query'
 import { toast } from 'sonner'
 import type { ResponseError, UseCustomMutationOptions } from 'types'
 
-import {
-  getScheduleDeleteCronJobRunDetailsKey,
-  getScheduleDeleteCronJobRunDetailsSql,
-} from './database-cron-jobs.sql'
+import { getScheduleDeleteCronJobRunDetailsKey } from './database-cron-jobs.utils'
 
 export type ScheduleCronJobRunDetailsCleanupVariables = {
   projectRef: string
