+++ apps/studio/data/database-extensions/database-extension-schema-query.ts
@@ -1,8 +1,8 @@
+import { getDatabaseExtensionDefaultSchemaSQL } from '@supabase/pg-meta'
 import { useQuery } from '@tanstack/react-query'
 import { UseCustomQueryOptions } from 'types'
 
 import { executeSql, ExecuteSqlError } from '../sql/execute-sql-query'
-import { getDatabaseExtensionDefaultSchemaSQL } from './database-extensions.sql'
 import { databaseExtensionsKeys } from './keys'
 
 type DatabaseExtensionDefaultSchemaVariables = {
